requirements.txt
praw==6.3.1
prawcore==1.0.1
psycopg2-binary==2.7.6.1